requirements.txt
pyautogen
langchain
openai
faiss-cpu
pypdf
langchain-community